agentic_engine/requirements.txt
kafka-python
redis
requests
pydantic
